requirements.txt
streamlit
stripe
pandas
plotly
python-dotenv
requests
fastapi
uvicorn
qrcode
pillow
reportlab
streamlit-webrtc
opencv-python-headless
beautifulsoup4>=4.9.0
lxml>=4.6.0
